added route protection for course_controller and started notes_controller

+++ app/controllers/notes_controller.rb
@@ -8,15 +8,25 @@ class NotesController < ApplicationController
   end
 
   def new
-    @note = Note.new
-    @courses = Course.all
+    if @current_user
+      @note = Note.new
+      @courses = Course.all
+    else 
+      redirect_to "/"
+      flash[:warning] = "You must be logged in to post."
+    end
   end
 
 
   def make_private
     n = Note.find(params[:id])
-    n.update(:whitelist => [@current_user.id.to_s])
-    redirect_to '/notes'
+    if @current_user.id == n.user.id
+      n.update(:whitelist => [@current_user.id.to_s])
+      redirect_to '/notes'
+    else 
+      redirect_to "/notes"
+      flash[:warning] = "You do not have the privilege for this."
+    end
   end
 
   def publish
@@ -132,11 +142,24 @@ class NotesController < ApplicationController
     redirect_to "/notes/#{n.id}"
   end
 
+
+# DOUBLE CHECK DELETION LOGIC - looks good.
   def destroy
-    NotesTags.where(:note_id => params[:id]).destroy_all
-    Note.find(params[:id]).delete
-    puts "request.referer" + request.referer
-    redirect_to '/courses'
+    n = Note.find(params[:id])
+    if (@current_user.id == n.user.id) || (@current_user.privilege == 2)
+      NotesTags.where(:note_id => params[:id]).destroy_all
+      c = Comment.where(:note_id => params[:id])
+      c.each do |child|
+        child.child_comments.destroy_all
+      end
+      c.destroy_all
+      n.delete
+      puts "request.referer" + request.referer
+      redirect_to '/courses'
+    else
+      redirect_to '/notes' 
+      flash[:danger] = "You cannot delete this note."
+    end
   end
 
   def edit
